# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Esta API corresponde al desarrollo del desafío tecnico a desarrollar en GO",
        "title": "API Operación Fuego de Quasar",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/topsecret/": {
            "post": {
                "description": "Obtains the position of the enemy ship and deciphers the message received on the satellites",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "TopSecret"
                ],
                "summary": "Get position and message",
                "parameters": [
                    {
                        "description": "Se debe enviar por el body la siguiente estructura JSON",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.TopSecret"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ResponseTopSecret"
                        }
                    }
                }
            }
        },
        "/topsecret_split/{satellite_name}": {
            "get": {
                "description": "Obtains the position of the enemy ship and deciphers the message received on the satellites",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "TopSecretSplit"
                ],
                "summary": "Get position and message for a specific satellite",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Nombre del Satélite",
                        "name": "satellite_name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ResponseTopSecret"
                        }
                    }
                }
            },
            "post": {
                "description": "Obtains the position of the enemy ship and deciphers the message received on the satellites",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "TopSecretSplit"
                ],
                "summary": "Get position and message for a specific satellite",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Nombre del Satélite",
                        "name": "satellite_name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Distancia de la nave enemiga",
                        "name": "Data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.TopSecret"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ResponseTopSecret"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.Entry": {
            "type": "object",
            "properties": {
                "distance": {
                    "type": "number"
                },
                "message": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "dto.Position": {
            "type": "object",
            "properties": {
                "x": {
                    "type": "number"
                },
                "y": {
                    "type": "number"
                }
            }
        },
        "dto.ResponseTopSecret": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "position": {
                    "$ref": "#/definitions/dto.Position"
                }
            }
        },
        "dto.TopSecret": {
            "type": "object",
            "properties": {
                "satellites": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.Entry"
                    }
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}